pyproject.toml
[project]
name = "grocery-delivery"
version = "0.1.0"
description = "An AI-powered system for recommending groceries to customers based on their grocery lists."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.4",
    "flask>=3.1.2",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.3",
]
